Initial commit from new directory
I started a new git repo in a new location. Could not find the old
(original) git repo, so I started a new one. This update contains an
updated RPM and Input Voltage guage, as well as a percentage guage for
the Testing function. RPM and Input Guages do not update with CAN
messages.